requirements.txt
django==4.2
environs==9.5.0
gunicorn==20.1.0
qrcode~=7.3.1
Pillow==9.5.0
phonenumbers==8.13.10
django-phonenumber-field[phonenumbers]==7.0.2
